pyproject.toml
[project]
name = "coolleddiy"
version = "0.1.0"
description = "CoolLed1248 16x64 LED-Matrix BLE Controller mit PySide6 GUI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PySide6>=6.6",
    "bleak>=0.21",
    "Pillow>=10.0",
    "qasync>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
